refactor: creates separate DateTime component
