+++ src/js/components/create/when.js
@@ -1,13 +1,9 @@
-
-/* eslint-disable no-shadow */
 import React, { Component } from 'react';
-import { View, Text, ScrollView, Switch } from 'react-native';
-import DatePicker from 'react-native-datepicker';
-import moment from 'moment';
-import Icon from 'react-native-vector-icons/FontAwesome';
+import { View, Text, ScrollView } from 'react-native';
 import Router from '../../router';
 import AddInput from '../general/add-input';
 import Button from '../common/Button';
+import DateTime from '../common/date-time';
 import styles from '../../../styles';
 import colours from '../../../styles/colours';
 
@@ -23,101 +19,13 @@ export default class When extends Component {
     }
   }
 
-  constructor () {
-    super();
-    this.state = {
-      tbcSwitch0: undefined,
-      tbcSwitch1: undefined,
-      tbcSwitch2: undefined
-    };
-  }
-
   nextPage = (name) => {
     this.props.navigator.push(Router.getRoute('confirm', { name }));
   };
 
   render () {
     const { name, data, addInput, handleDate, handleTime, removeInput } = this.props;
-    const inputs = data.map((value, i) => {
-      return (
-        <View key={ Math.random() }>
-          <View style={{ margin: 10 }}>
-            <DatePicker
-              style={{ width: 200 }}
-              date={ value.date }
-              mode="date"
-              placeholder="select date"
-              format="DD/MM/YYYY"
-              minDate={ moment().format('DD MM YYYY')}
-              confirmBtnText="Confirm"
-              cancelBtnText="Cancel"
-              customStyles={{
-                dateIcon: {
-                  position: 'absolute',
-                  left: 0,
-                  top: 4,
-                  marginLeft: 0
-                },
-                dateInput: {
-                  marginLeft: 36
-                }
-              }}
-              onDateChange={ date => handleDate(date, i) }
-            />
-          </View>
 
-          <View style={{ margin: 10 }}>
-            <DatePicker
-              style={{ width: 200 }}
-              date={ value.time && value.time }
-              mode="time"
-              placeholder="select time"
-              format="HH:mm"
-              confirmBtnText="Confirm"
-              cancelBtnText="Cancel"
-              minuteInterval={10}
-              customStyles={{
-                dateIcon: {
-                  position: 'absolute',
-                  left: 0,
-                  top: 4,
-                  marginLeft: 0
-                },
-                dateInput: {
-                  marginLeft: 36
-                }
-              }}
-              onDateChange={ (time) => {
-                this.setState({ [`tbcSwitch${i}`]: false });
-                handleTime(time, i);
-              }}
-            />
-          </View>
-          { i !== 0 &&
-            <Icon
-              name="remove"
-              size={18}
-              color="gray"
-              style={{ position: 'absolute', right: 100, top: 20 }}
-              onPress={ removeInput }
-            />
-          }
-          <View style={{ flexDirection: 'row', position: 'absolute', right: 40, top: 75 }}>
-            <Switch
-              onValueChange={ (switchValue) => {
-                this.setState({ [`tbcSwitch${i}`]: switchValue });
-                const newTimeValue = switchValue ? '' : moment().format('HH:mm');
-                handleTime(newTimeValue, i);
-              }}
-
-              value={ this.state[`tbcSwitch${i}`] }
-            />
-            <Text style={{ margin: 5 }}>TBC</Text>
-          </View>
-
-        </View>
-      );
-    });
 
     const hideNext = data[0].date === '';
 
@@ -133,7 +41,12 @@ export default class When extends Component {
           </Text>
         </View>
         <View style={styles.whenContainer}>
-          { inputs }
+          <DateTime
+            data={data}
+            handleDate={handleDate}
+            handleTime={handleTime}
+            removeInput={removeInput}
+          />
           <AddInput data={ data } handler={ addInput } />
         </View>
         <View style={styles.container}>
